Assignment 42: complete

+++ app/controllers/comments_controller.rb
@@ -3,31 +3,59 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
     else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to @topic
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    comment = Comment.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted."
-      redirect_to [@post.topic, @post]
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
     else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      @topic = Topic.find(params[:topic_id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to @topic
+      end
     end
   end
+
 
   private
